GNUmakefile: Remove -Werror
Compiler errors change all the time (across compiler implementations,
versions, and platforms). Thus, using it as part of the standard build
process almost guarantees that the project will become unbuildable in
certain conditions. While -Werror makes sense to use on developer
machines and CI, it should not be part of the build process as used to
compile the distributed source code.

Fixes #9.